Fixing a couple of bugs I introduced.
1. The switch statement for the enum had an unquoted "Default", which meant it was the default handler for the switch, NOT matching the literal string 'default'.  This caused a mysterious Windows-1252 to appear.
2. Get-UniCodepoint just needs the WebNames of the encodings and not the full Encoding objects.

I also fixed the switch statement formatting to match the pattern in use elsewhere, as well as made it no longer require a cast to the enum.
This allowed me to add support for 'UTF16' (no hyphen), which isn't part of the enum, but seems like it's an omission if you support 'UTF8' but not 'UTF16'.